Warnings cleanup (#49)
* fix: sign conversion in _frameNumber

* fix: unused parameter in mesh getter

* fix: double-float conversion in updateCube

* fix: removed -Wmissing-field-initializers flag

* fix: float conversion of fileSize in vk_pipelines.cpp

* fix: added array indexing to silence string plus int warning

* fix: unused parameters

* fix: signedness changing conversions fix in vk_images.cpp

* fix: signedness changing and float conversions fix in vk_descriptors.cpp.cpp

* fix: changed field initialization order in Node ctor

* fix: unused parameter in TransformSystem.cpp

* fix: float-double conversion

* fix: range-based for loops

* fix: uint32_t and size_t conversions

* fix: capture in lambda function for deletionQueue

* fix: sign and long conversion

* fix: removed range based for loop

* fix: removed DescriptorWriter variable re-declaring

* fix: long conversion

* fix: added default ctor in IRenderable heirs

* fix(vk_pipelines.cpp): casting fileSize to uint32_t

* fix: int parameter to string conversion in "cube" + parameter

* fix: fileSize.tellg() error handling

* fix: removed unnecessary std::move in Cubes functions

+++ src/graphics/vulkan/vk_images.cpp
@@ -41,12 +41,12 @@ void vkutil::copy_image_to_image(VkCommandBuffer cmd, VkImage source,
     VkImageBlit2 blitRegion{.sType = VK_STRUCTURE_TYPE_IMAGE_BLIT_2,
                             .pNext = nullptr};
 
-    blitRegion.srcOffsets[1].x = srcSize.width;
-    blitRegion.srcOffsets[1].y = srcSize.height;
+    blitRegion.srcOffsets[1].x = static_cast<int32_t>(srcSize.width);
+    blitRegion.srcOffsets[1].y = static_cast<int32_t>(srcSize.height);
     blitRegion.srcOffsets[1].z = 1;
 
-    blitRegion.dstOffsets[1].x = dstSize.width;
-    blitRegion.dstOffsets[1].y = dstSize.height;
+    blitRegion.dstOffsets[1].x = static_cast<int32_t>(dstSize.width);
+    blitRegion.dstOffsets[1].y = static_cast<int32_t>(dstSize.height);
     blitRegion.dstOffsets[1].z = 1;
 
     blitRegion.srcSubresource.aspectMask = VK_IMAGE_ASPECT_COLOR_BIT;
